Alot of changes like bug fix and credit

+++ Assets/Script/DropTable/DropItem.cs
@@ -8,22 +8,26 @@ public class DropItem : MonoBehaviour
     [SerializeField]private DropTable dropTable;
     [SerializeField]private GameObject itemPrefab;
     [SerializeField]private List<ItemData> DropList;
-
+    private bool IsDropping = false;
     public void ItemDrop(Transform transform)
     {
-        Table table = dropTable.LootingTable[0];
-        DropList = dropTable.RollLoot(table);
-        foreach (ItemData item in DropList)
+        if (!IsDropping)
         {
-            GameObject obj = Instantiate(itemPrefab, transform.position, Quaternion.identity);
-            obj.GetComponent<WorldItem>().SetItem(item,1);
-            Rigidbody2D rb = obj.AddComponent<Rigidbody2D>();
-            rb.constraints = RigidbodyConstraints2D.FreezeRotation;
-            if (rb != null)
-                StartCoroutine(ApplyFloatDrop(rb));
+            IsDropping=true;
+            Table table = dropTable.LootingTable[0];
+            DropList = dropTable.RollLoot(table);
 
+            foreach (ItemData item in DropList)
+            {
+                GameObject obj = Instantiate(itemPrefab, transform.position, Quaternion.identity);
+                obj.GetComponent<WorldItem>().SetItem(item, 1);
+                Rigidbody2D rb = obj.AddComponent<Rigidbody2D>();
+                rb.constraints = RigidbodyConstraints2D.FreezeRotation;
+                if (rb != null)
+                    StartCoroutine(ApplyFloatDrop(rb));
+
+            }
         }
-     
     }
 
     public void DropSpecficItem(ItemData item)
